# react-frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Furballs Backend API
  description: E-commerce backend API built with FastAPI for the Keychain Crew (Furballs) online store
  version: 1.0.0
  contact:
    name: Keychain Crew Development Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.soar-commerce.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - email
        - first_name
        - last_name
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        first_name:
          type: string
        last_name:
          type: string
      required:
        - email
        - password
        - first_name
        - last_name
...
    Image:
      type: object
      properties:
        image_url:
          type: string
          format: uri
        alt_text:
          type: string
        image_priority:
          type: integer
          minimum: 0
      required:
        - image_url
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
        price:
          type: number
          minimum: 0
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        item_number:
          type: string
        category:
          type: string
        stock:
          type: integer
          minimum: 0
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - name
        - description
        - price
        - category
    ProductCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
        price:
          type: number
          minimum: 0
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        item_number:
          type: string
        category:
          type: string
        stock:
          type: integer
          minimum: 0
      required:
        - name
        - description
        - price
        - category
...
    Error:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - auth
      summary: Register New User
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/products:
    get:
      tags:
        - products
      summary: List Products
      description: List all products with pagination
      parameters:
        - name: id
          in: query
          description: Product ID to get single product
          schema:
            type: string
        - name: skip
          in: query
          description: Number of products to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Maximum number of products to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      tags:
        - products
      summary: Create Product
      description: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
  /api/v1/products/{product_id}:
    get:
      tags:
        - products
      summary: Get Product
      description: Get a single product by ID
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - products
      summary: Update Product
      description: Update an existing product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '200':
          description: Product updated successfully
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []

    delete:
      tags:
        - products
      summary: Delete Product
      description: Delete a product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
...
  /api/v1/checkout:
    post:
      tags:
        - checkout
      summary: Create Checkout Session
      description: Create checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: string
                      quantity:
                        type: integer
                        minimum: 1
              required:
                - items
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  checkout_url:
                    type: string
                    format: uri
                  session_id:
                    type: string
      security:
        - BearerAuth: []
...